pyproject.toml
[project]
name = "llmstxt"
version = "0.1.0"
description = "Generate an llmtxt file for a git repository"
readme = "README.md"
authors = [
    { name = "Terence Munro", email = "terence.munro@microba.com" }
]
requires-python = ">=3.13"
dependencies = [
  "llama-index==0.10.10",
  "llama-index-llms-openai==0.1.5",
  "llama-index-readers-file==0.1.4",
  "gitpython==3.1.41",
  "python-dotenv==1.0.0",
  "openai==1.12.0",
]

[project.scripts]
llmstxt = "llmstxt.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pyright>=1.1.401",
    "ruff>=0.11.11",
]
